requirements.txt
# 必要なライブラリを記述したファイル

Flask==3.0.3                   # Webアプリケーションフレームワーク
gunicorn==23.0.0               # 作成したバックエンドを実行するWebサーバの機能を提供するライブラリ
google-cloud-bigquery==3.25.0  # BigQueryを操作するためのライブラリ